fix: relacao entre user e role

+++ Core/IBGEExplorer.Account/Entities/Role.cs
@@ -5,7 +5,7 @@ namespace IBGEExplorer.Account.Entities;
 public class Role : Entity<short>
 {
     public string? Name { get; set; }
-    public int RoleId { get; set; }
+    public short RoleId { get; set; }
 
     public List<UserRole>? UserRoles { get; set; }
 }
